+++ app/src/main/java/com/example/daoandroid/database/models/CgMass.java
@@ -3,7 +3,7 @@ package com.example.daoandroid.database.models;
 public class CgMass {
 
     //compose cg id_model/nameModel
-    private int cg;
+    private double cg;
     private String nameModel;
 
     //max mass permit
@@ -13,16 +13,16 @@ public class CgMass {
     public CgMass(String nameModel) {
         this.nameModel = nameModel;
     }
-    public CgMass(int cg, String nameModel, double mass) {
+    public CgMass(double cg, String nameModel, double mass) {
         this.cg = cg;
         this.nameModel = nameModel;
         this.mass = mass;
     }
 
-    public int getCg() {
+    public double getCg() {
         return cg;
     }
-    public void setCg(int cg) {
+    public void setCg(double cg) {
         this.cg = cg;
     }
     public String getNameModel() {
